:tada: init: Add AstPrinter class for printing Abstract Syntax Trees (ASTs)
Initialized the project and added the AstPrinter class, which implements the Expr.Visitor interface to print Abstract Syntax Trees (ASTs) of expressions. The AstPrinter class defines methods to visit different types of expression nodes, such as binary expressions, grouping expressions, literal expressions, and unary expressions. It provides a print method to generate a string representation of the AST for a given expression. Additionally, a main method is included to demonstrate the usage of the AstPrinter class by printing an example expression.